Handle unpowered devices
Previously, this logic would error when attempting to operate on an
unpowered device because the method/interface wouldn't exist.
Unfortunately, there's no race-free way to check if the device is
powered before operating on it, so I just ignore errors.

I'm also adding a "powered" status to devices returned by
`iwd--get-devices'. It's not currently used, but it doesn't cost us
anything and I expect we'll need it later.